Allow core data fetch capability
Allow user to select favourite city and view saved forecast

+++ Forecaster/CityForecast+CoreDataProperties.swift
@@ -16,7 +16,7 @@ extension CityForecast {
         return NSFetchRequest<CityForecast>(entityName: "CityForecast")
     }
 
-    @NSManaged public var dayOfWeek: Int16
+    @NSManaged public var dayOfWeek: Int32
     @NSManaged public var currentTemperature: Double
     @NSManaged public var condition: Int16
     @NSManaged public var relationship: FavouriteCity
